+++ src/import_roms/test_chd_multiple_tracks.rs
@@ -52,8 +52,7 @@ async fn test() {
     import_chd(
         &mut connection,
         &progress_bar,
-        &system_directory,
-        &system,
+        Some(&system),
         &None,
         &romfile_path,
         &HashAlgorithm::Crc,