+++ src/import_roms/test_cso.rs
@@ -54,8 +54,7 @@ async fn test() {
     import_cso(
         &mut connection,
         &progress_bar,
-        &system_directory,
-        &system,
+        Some(&system),
         &None,
         &romfile_path,
         &HashAlgorithm::Crc,