+++ src/import_roms/test_sevenzip_multiple_files_full_game.rs
@@ -45,8 +45,7 @@ async fn test() {
     import_archive(
         &mut connection,
         &progress_bar,
-        &system_directory,
-        &system,
+        Some(&system),
         &None,
         &romfile_path,
         romfile_path.extension().unwrap().to_str().unwrap(),